refactor(entities): move entities to src/entities folder
