+++ src/entities/BlockNode/types/BlockNodeConstructorParameters.ts
