+++ src/entities/BlockNode/types/index.ts
