+++ src/entities/EditorDocument/types/EditorDocumentConstructorParameters.ts
