+++ src/entities/EditorDocument/types/PropName.ts
