+++ src/entities/EditorDocument/types/index.ts
