+++ src/entities/TextNode/index.ts
