+++ src/entities/TextNode/types/TextNodeConstructorParameters.ts
