+++ src/entities/TextNode/types/index.ts
